% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_avg.R
\name{bat_avg}
\alias{bat_avg}
\title{Calculate batting average of batsmen by season}
\usage{
bat_avg(players, years)
}
\arguments{
\item{players}{a list of batsmen for whom batting average has to be calculated,
as character vectors}

\item{years}{a list of seasons (years) for which the batting average of the
given batsmen has to be calculated, as numeric vectors}
}
\description{
Calculate batting average of batsmen by season
}
\examples{

library(ipl)

# Compute batting average of Virat Kohli in 2019
bat_avg("V Kohli", 2019)


# Compute batting averages of Virat Kohli, MS Dhoni, Rohit Sharma in
# 2017, 2018, and 2019
bat_avg(c("V Kohli", "MS Dhoni", "RG Sharma"), c(2017, 2018, 2019))






}
