% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bat_max.R
\name{bat_max}
\alias{bat_max}
\title{Calculate the maximum runs made by a batsman in a given season}
\usage{
bat_max(player, yr)
}
\arguments{
\item{player}{the batsman for whom maximum runs have to be calculated, as a character vector}

\item{yr}{the season for which the maximum runs of the given batsman have to be calculated, as
a numeric vector}
}
\description{
Calculate the maximum runs made by a batsman in a given season
}
\examples{

library(ipl)

# Find maximum runs made by Virat Kohli in the year 2016
bat_max("V Kohli", 2016)

# Find maximum runs made by AB de Villiers in the year 2019
bat_max("AB de Villiers", 2019)








}
