% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{batsman_100}
\alias{batsman_100}
\title{This file contains all the details of the top 100 batsmen}
\format{
A data frame of 100 rows representing the stats of the top 100 batsman
\describe{
\item{pos}{The position of the player in the 1 to 100 ranking}
\item{player}{Player name}
\item{mat}{Number of matches played by the batsman}
\item{inns}{The number of innings played by the batsman}
\item{no}{Number of times the batsman was not out at the end of an inning}
\item{runs}{The total number of runs made by the batsman}
\item{hs}{Highest score ever made by the batsman}
\item{avg}{Average batting score}
\item{bf}{Number of balls faced by the batsman}
\item{sr}{The overall strike rate of the batsman}
\item{num_100}{The number of innings where the batsman scored one hundred or more runs}
\item{num_50}{The number of innings where the batsman scored fifty to ninety runs}
\item{num_4s}{Number of 4s scored by the batsman}
\item{num_6s}{Number of 6s scored by the batsman}
}
}
\source{
This dataset was obtained from Kaggle: \href{https://www.kaggle.com/datasets/nkitgupta/ipl-auction-and-ipl-dataset?select=Top_100_bowlers.xlsx}{Top_100_Batsman}.
}
\usage{
batsman_100
}
\description{
This file contains all the details of the top 100 batsmen
}
\keyword{datasets}
