% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batsman_summary.R
\name{batsman_summary}
\alias{batsman_summary}
\title{Get the summary table of all batting statistics of a batsman by year
(across all IPL matches between 2008 and 2020)}
\usage{
batsman_summary(player)
}
\arguments{
\item{player}{The name of the batsman for whom the batting summary table is
desired}
}
\value{
\code{batsman_summary} returns the a data frame with all batting statistics
of the batsman for all years
}
\description{
Get the summary table of all batting statistics of a batsman by year
(across all IPL matches between 2008 and 2020)
}
\examples{

library(ipl)

# Get the summary table of all batting statistics for MS Dhoni
batsman_summary("MS Dhoni")
}
