% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowler_names.R
\name{bowler_names}
\alias{bowler_names}
\title{Returns all the bowlers with an average score higher than the given score}
\usage{
bowler_names(given_score)
}
\arguments{
\item{given_score}{The score for which we want all the bowler's names for, as a numeric vector}
}
\value{
\code{bowler_names} returns a list of names.
}
\description{
Returns all the bowlers with an average score higher than the given score
}
\examples{

library(ipl)

# Finds all the bowlers with a score higher than 30
bowler_names(30)
}
