% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowler_score.R
\name{bowler_score}
\alias{bowler_score}
\title{Returns all the bowlers with an average score higher than the given score}
\usage{
bowler_score(given_score)
}
\arguments{
\item{given_score}{The avg score for which we want all the bowler's names for, as a numeric vector}
}
\value{
\code{bowler_score} returns a list of names.
}
\description{
Returns all the bowlers with an average score higher than the given score
}
\examples{

library(ipl)

# Finds all the bowlers with an average score higher than 30
bowler_score(30)
}
