% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowler_summary.R
\name{bowler_summary}
\alias{bowler_summary}
\title{Get the summary table of the bowling statistics of a batsman
(across all IPL matches between 2008 and 2020)}
\usage{
bowler_summary()
}
\value{
\code{bowler_summary} returns the a data frame with the bowling statistics
}
\description{
Get the summary table of the bowling statistics of a batsman
(across all IPL matches between 2008 and 2020)
}
\examples{

library(ipl)

# Get the summary table of the bowler statistics for Rahul Sharma
bowler_summary("Rahul Sharma")
}
