% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowlers_100}
\alias{bowlers_100}
\title{Information on top 100 bowlers of IPL}
\format{
A data frame of 100 rows for all the bowlers
\describe{
\item{pos}{The position of the player in the 1 to 100 ranking}
\item{player}{Player name}
\item{mat}{Number of matches played by bowler}
\item{inns}{The number of innings}
\item{ov}{The number of overs}
\item{runs}{The number of runs}
\item{wkts}{The number of wickets}
\item{bbi}{Best Bowling in Innings (BBI) gives the score for one inning}
\item{avg}{The average bowling score for the player}
\item{econ}{bowler economy}
\item{sr}{The strike rate of the bowler}
\item{num_4w}{The number of innings in which the bowler took exactly four wickets}
\item{num_5w}{the number of innings in which the bowler took exactly five wickets}
}
}
\source{
This dataset was obtained from Kaggle: \href{https://www.kaggle.com/datasets/nkitgupta/ipl-auction-and-ipl-dataset}{Bowlers}.
}
\usage{
bowlers_100
}
\description{
Information on top 100 bowlers of IPL
}
\keyword{datasets}
