% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowling_analysis.R
\name{bowling_analysis}
\alias{bowling_analysis}
\title{Returns the summary analysis for the given bowler}
\usage{
bowling_analysis(player_name)
}
\arguments{
\item{player_name}{The bowler whose summary analysis has to be calculated, as a
character vector}
}
\value{
\code{bowling_analysis} returns the summary analysis of a bowler in a data frame
}
\description{
Returns the summary analysis for the given bowler
}
\examples{

library(ipl)

# Summarises the bowling analysis for Rahul Sharma
bowling_analysis("Rahul Sharma")
}
