% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cents_halfcents.R
\name{cents_halfcents}
\alias{cents_halfcents}
\title{Calculate the number of centuries and half-centuries made a batsman in a given
season}
\usage{
cents_halfcents(player, yr)
}
\arguments{
\item{player}{The batsman for whom the number of centuries and half-centuries
have to be calculated, as a character vectors}

\item{yr}{The season (year) for which the number of centuries and
half-centuries for the given batsman have to be calculated, as a numeric vector}
}
\description{
Calculate the number of centuries and half-centuries made a batsman in a given
season
}
\examples{

library(ipl)

# Calculate the number of centuries and half-centuries made by Virat Kohli
# in the 2016 IPL season

cents_halfcents("V Kohli", 2016)
}
