% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fours.R
\name{fours}
\alias{fours}
\title{Find a batsman's total number of fours}
\usage{
fours(player_name)
}
\arguments{
\item{player_name}{a batsman for whom we want to calculate the number of fours they scored}
}
\value{
\code{sixes} return number of fours for the given player as a value
}
\description{
Find a batsman's total number of fours
}
\examples{
Compute the total number of fours scored by Suresh Raina
fours("Suresh Raina")

library(ipl)

}
