% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ipl}
\alias{ipl}
\title{Information on matches from 2008 to 2020}
\format{
A data frame of 816 rows containing the information about all matches from 2008 to 2020
\describe{
\item{id}{Unique match ID of the IPL match as per ESPNCricinfo}
\item{city}{City in which the match took place}
\item{date}{Date of the match}
\item{year}{Year (season) of IPL match}
\item{player_of_match}{Recipient of the player of match award}
\item{venue}{Venue where the match took place}
\item{neutral_venue}{Neutral }
\item{team1}{First team name}
\item{team2}{Second team name}
\item{toss_winner}{Winner of the coin toss}
\item{toss_decision}{Whether toss winning team chose to bat or field first}
\item{winner}{Winning team}
\item{result}{Wickets or Runs}
\item{result_margin}{Margin won by}
\item{eliminator}{Y/N advanced to The Humdred Eliminator}
\item{method}{Method used to calculate the target score}
\item{umpire1}{Name of first umpire}
\item{umpire2}{Name of second umpire}
}
}
\source{
This dataset was obtained from Kaggle: \href{https://www.kaggle.com/datasets/patrickb1912/ipl-complete-dataset-20082020?select=IPL+Matches+2008-2020.csv}{IPL Matches 2008-2020}
}
\usage{
ipl
}
\description{
Information on matches from 2008 to 2020
}
\keyword{datasets}
