% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the IPL Auction Shiny App}
\usage{
launch_app()
}
\value{
Launches the interactive Shiny app in a browser window.
The app allows users to filter players by team and performance metrics.
}
\description{
This function launches the Shiny app for exploring IPL 2022 player performance data.
}
\examples{
# To launch the Shiny app:
launch_app()
}
