% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maiden_overs.R
\name{maiden_overs}
\alias{maiden_overs}
\title{Returns the maiden overs for the given bowler}
\usage{
maiden_overs(player_name)
}
\arguments{
\item{player_name}{The bowler whose maiden overs has to be calculated, as a
character vector}
}
\value{
\code{maidenOvers} returns the maiden overs
}
\description{
Returns the maiden overs for the given bowler
}
\examples{

library(ipl)

# Calculates the maiden overs for Rahul Sharma
maiden_overs("Rahul Sharma")
}
