% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overs.R
\name{overs}
\alias{overs}
\title{Calculates the number of overs for the given bowler}
\usage{
overs(player)
}
\arguments{
\item{player}{The bowler whose overs has to be calculated, as a character vector}
}
\value{
\code{overs} returns the number of overs
}
\description{
Calculates the number of overs for the given bowler
}
\examples{

library(ipl)

# Calculates the number of overs for Rahul Sharma
overs("Rahul Sharma")
}
