% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overs_balls.R
\name{overs_balls}
\alias{overs_balls}
\title{Calculates the completed overs and remaining (legal) balls bowled by
the given bowler across all IPL matches from 2008-2020.}
\usage{
overs_balls(player_name)
}
\arguments{
\item{player_name}{The bowler whose completed overs and remaining balls has to
be calculated, as a character vector}
}
\value{
\code{overs_balls} returns the number of completed overs and remaining balls
bowled by the given player across all IPL matches from 2008-2020
}
\description{
Calculates the completed overs and remaining (legal) balls bowled by
the given bowler across all IPL matches from 2008-2020.
}
\examples{

library(ipl)

# Calculates the completed runs and remaining (legal) balls bowled by Rahul Sharma
overs_balls("Rahul Sharma")
}
