% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partnership_runs.R
\name{partnership_runs}
\alias{partnership_runs}
\title{Calculate the runs made by all batting partnerships in a given IPL match for
a given team}
\usage{
partnership_runs(match_id, batting_side)
}
\arguments{
\item{match_id}{The unique match id for the desired IPL match, according to
ESPNCricInfo}

\item{batting_side}{The particular batting team in the given IPL match for whom
partenrship runs have to be calculated}
}
\value{
\code{partnership_runs} returns a stacked bar plot with the batting
partnerships on the x-axis, the total runs made by the partnership on the
y-axis, and the bars filled in by the number of runs made by each batsman in the
partnership
}
\description{
Calculate the runs made by all batting partnerships in a given IPL match for
a given team
}
\examples{

library(ipl)

# Calculate the number of runs made by Mumbai Indians in their match against
Delhi Capitals on 2019-03-24, by partnerships
partnership_runs(1175358, "Mumbai Indians")

}
