% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs.R
\name{runs}
\alias{runs}
\title{Calculates the number of runs for the given bowler}
\usage{
runs(player_name)
}
\arguments{
\item{player_name}{The bowler whose runs has to be calculated, as a character vector}
}
\value{
\code{runs} returns the number of runs conceded by the given bowler
}
\description{
Calculates the number of runs for the given bowler
}
\examples{

library(ipl)

# Calculates the number of runs for Rahul Sharma
runs("Rahul Sharma")
}
