% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Define the server-side logic for the Shiny app}
\usage{
server(input, output)
}
\arguments{
\item{input}{The input object from the UI}

\item{output}{The output object that renders the scatter plot}
}
\value{
The server logic that updates the scatter plot based on user inputs

Generates scatter plot of batting or bowling metrics
}
\description{
This function handles the user inputs (player type and team filters),
filters the IPL player data, and generates a scatter plot
based on the filtered data.
}
