% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sixes.R
\name{sixes}
\alias{sixes}
\title{Find a batsman's total number of sixes}
\usage{
sixes(player_name, yr)
}
\arguments{
\item{player_name}{a batsman for whom we want to calculate the number of sixes they scored}
}
\value{
\code{sixes} return number of sixes for the given player as a value
}
\description{
Find a batsman's total number of sixes
}
\examples{
library(ipl)

# Find the total number of sixes scored by Rohit Sharma in the 2017 IPL season
sixes("RG Sharma", 2017)
}
