% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowler_score.R, R/strike_rate.R
\name{strike_rate}
\alias{strike_rate}
\title{Returns all the bowlers with an average score higher than the given score}
\usage{
strike_rate(player, yr)

strike_rate(player, yr)
}
\arguments{
\item{player}{The batsman whose strike rate has to be calculated, as a character vector}

\item{yr}{The year for which the strike rate of the given batsman has to be calculated,
as a numeric vector}

\item{score}{The score for which we want all the bowler's names for, as a numeric vector}
}
\value{
\code{bowlers_list} returns a list of names.

\code{strike_rate} returns a tibble with one row and two columns.
}
\description{
Returns all the bowlers with an average score higher than the given score

Calculate strike rate of a given batsman in a given season
}
\examples{

library(ipl)

# Finds all the bowlers with a score higher than 30
bowler_names(30)

library(ipl)

# Calculate the strike rate of MS Dhoni in the 2016 IPL season
strike_rate("MS Dhoni", 2016)
}
