% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Define the user interface (UI) for the Shiny app}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
A fluid page layout for the Shiny app UI
}
\description{
This UI allows users to select a player type (batsman or bowler),
choose a team, and visualize performance metrics in a scatter plot.
}
\keyword{datasets}
