% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wickets_taken.R
\name{wickets_taken}
\alias{wickets_taken}
\title{This function returns the number of wickets taken for the player}
\usage{
wickets_taken(player_name)
}
\arguments{
\item{player_name}{The name of the given player for whom we want to calculate
the number of wickets taken}
}
\value{
\code{num_wickets} number of wickets taken by the given bowler
}
\description{
This function returns the number of wickets taken for the player
}
\examples{

library(ipl)

# Returns the number of wickets taken by Rahul Sharma
wickets_taken("Rahul Sharma")
}
