% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winloss.R
\name{winloss}
\alias{winloss}
\title{Calculates the wins, losses, and win percentage for a team based on their starting position (batting or fielding)}
\usage{
winloss(team_name, year)
}
\arguments{
\item{team_name}{a team for whom number of wins, losses, and win percentage has to be calculated}

\item{year}{a list of seasons (years) for which number of wins, losses, and win percentage of the
given team has to be calculated}
}
\value{
\code{winloss} returns a tibble with 3 columns and 3 rows
}
\description{
Calculates the wins, losses, and win percentage for a team based on their starting position (batting or fielding)
}
\examples{
Calculates the wins, losses, and win percentage for Sunrisers Hyderabad in 2017
winloss("Sunrisers Hyderabad", 2017)

library(ipl)

}
