% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bowlers}
\alias{bowlers}
\title{This file contains all the details of the top 100 bowlers}
\format{
A data frame of 100 rows for all the bowlers
\describe{
\item{pos}{The position of the player in the 1 to 100 ranking}
\item{player}{Player name}
\item{mat}{ }
\item{inns}{The number of innings}
\item{ov}{The number of overs}
\item{runs}{The number of runs}
\item{wks}{The number of wickets}
\item{bbi}{Best Bowling in Innings (BBI) gives the score for one inning}
\item{avg}{ }
\item{econ}{ }
}
}
\usage{
bowlers
}
\description{
This file contains all the details of the top 100 bowlers
}
\keyword{datasets}
