% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{deliveries}
\alias{deliveries}
\title{Ball-by-ball data of IPL matches played in 2008-2020}
\format{
A data frame of 193468 rows representing each ball bowled/ played in
all IPL matches from 2008-2020
}
\usage{
deliveries
}
\description{
Ball-by-ball data of IPL matches played in 2008-2020
}
\keyword{datasets}
