% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strike_rate.R
\name{strike_rate}
\alias{strike_rate}
\title{Calculate strike rate of a given batsman in a given season}
\usage{
strike_rate(player, yr)
}
\arguments{
\item{player}{The batsman whose strike rate has to be calculated, as a character vector}

\item{yr}{The year for which the strike rate of the given batsman has to be calculated,
as a numeric vector}
}
\description{
Calculate strike rate of a given batsman in a given season
}
\examples{

library(ipl)

 # Calculate the strike rate of MS Dhoni in the 2016 IPL season
strike_rate("MS Dhoni", 2016)






}
