% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{teams}
\alias{teams}
\title{Winning team, overs bowled, runs made and wickets fallen for each team in
each IPL match played in 2008-2020}
\format{
A data frame of 1630 rows representing the result, runs made, overs played, and wickets
fallen for each team in each IPL match played in 2008-2020
}
\usage{
teams
}
\description{
Winning team, overs bowled, runs made and wickets fallen for each team in
each IPL match played in 2008-2020
}
\keyword{datasets}
