% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss_choice.R
\name{toss_choice}
\alias{toss_choice}
\title{The number of times a team chooses to start batting and fielding}
\usage{
toss_choice(team_name)
}
\arguments{
\item{team_name}{a team for whom number batting versus fielding statistics have to be calculated}

\item{year}{a list of seasons (years) for which the statistics of the given team have to be calculated}
}
\value{
\code{toss_choice} returns a tibble with 2 columns and 1 rows
}
\description{
The number of times a team chooses to start batting and fielding
}
\examples{
library(ipl)

# The number of times Dehli Daredevils chooses to start batting and fielding
toss_choice("Delhi Daredevils")


}
