% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winloss.R
\name{winloss}
\alias{winloss}
\title{Get a summary table of a team's wins and losses for a given year}
\usage{
winloss(team_name, yr)
}
\arguments{
\item{team_name}{A team for whom we want to calculate the number of wins and losses}

\item{yr}{The year for which the number of wins and losses for the given team is to be
calculated}
}
\value{
\code{winloss} return number of wins, losses, and win percentage at each starting position
for the given team on the given year, as a data frame
}
\description{
Get a summary table of a team's wins and losses for a given year
}
\examples{
library(ipl)

# Find the total number of wins and losses for Sunrisers Hyderabad in the 2017 IPL season
winloss("Delhi Daredevils", 2017)
}
